# backend/src/modules/alerting/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: TelemetryFlow Alerting API
  description: Alerting & Notification module API
  version: 1.0.0
servers:
  - url: /api/v2
...
paths:
...
  /alert-instances:
    get:
      tags: [Alert Instances]
      summary: List alert instances
      operationId: listAlertInstances
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [pending, firing, resolved]
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Instance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertInstance"
  /alert-instances/{id}/acknowledge:
    post:
      tags: [Alert Instances]
      summary: Acknowledge alert
      operationId: acknowledgeAlert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Alert acknowledged
  /alert-instances/{id}/resolve:
    post:
      tags: [Alert Instances]
      summary: Resolve alert
      operationId: resolveAlert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Alert resolved
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AlertInstance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ruleId:
          type: string
          format: uuid
        ruleName:
          type: string
        state:
          type: string
          enum: [pending, firing, resolved]
        severity:
          type: string
        labels:
          type: object
        annotations:
          type: object
        value:
          type: number
        startedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
...